+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/lib/arguments/PlayerByUUIDArgument.java
@@ -27,12 +27,12 @@ public class PlayerByUUIDArgument extends ObjectByIdArgument<UUID, Player> {
 	}
 
 	public PlayerByUUIDArgument(String name, ArgumentFilter<Player> filter, int minimalCompletionInput) {
-		super(name, filter, minimalCompletionInput);
+		super(name, filter, new IdArgumentArgs(minimalCompletionInput));
 	}
 
 	@Override
-	protected ObjectIdArgument<UUID> createIdArgument(String name, int minimalCompletionInput) {
-		return new PlayerUUIDArgument(name, ArgumentFilter.acceptAny(), minimalCompletionInput) {
+	protected ObjectIdArgument<UUID> createIdArgument(String name, IdArgumentArgs args) {
+		return new PlayerUUIDArgument(name, ArgumentFilter.acceptAny(), args.minimalCompletionInput) {
 			@Override
 			protected Iterable<UUID> getCompletionSuggestions(CommandInput input, CommandContextView context, String idPrefix) {
 				return PlayerByUUIDArgument.this.getCompletionSuggestions(input, context, idPrefix);